requirements.txt
argcomplete==1.12.2
boto3==1.16.63
botocore==1.19.63
cachetools==4.2.0
certifi==2020.12.5
cfn-flip==1.2.3
chardet==4.0.0
click==7.1.2
durationpy==0.5
Flask==1.1.2
future==0.18.2
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gspread==3.6.0
hjson==3.0.2
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kappa==0.6.0
MarkupSafe==1.1.1
numpy==1.20.0
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.9
pandas==1.2.1
pip-tools==5.5.0
placebo==0.9.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
s3transfer==0.3.4
six==1.15.0
slack-sdk==3.2.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.56.0
troposphere==2.6.3
uritemplate==3.0.1
urllib3==1.26.2
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.52.0
